+++ src/enum-collection.ts
@@ -1,20 +1,17 @@
 import type { EnumExtension } from 'enum-plus/extension';
-import type { EnumItemOptions } from './enum-item';
-import { EnumItemClass } from './enum-item';
+import type { EnumItemClass, EnumItemOptions } from './enum-item';
 import type { IEnumItems, ToListConfig } from './enum-items';
 import { EnumItemsArray } from './enum-items';
-import { localizer } from './localize';
+import { localizer } from './localizer';
 import type {
   ColumnFilterItem,
   EnumInit,
-  EnumItemInit,
   EnumKey,
   EnumValue,
   FindEnumKeyByValue,
   ListItem,
   MenuItemOption,
   PrimitiveOf,
-  StandardEnumItemInit,
   ValueTypeFromSingleInit,
 } from './types';
 import type { IS_ENUM_ITEMS } from './utils';
@@ -42,11 +39,50 @@ export class EnumCollectionClass<
     V extends EnumValue = ValueTypeFromSingleInit<T[K], K>,
   >
   extends EnumExtensionClass<T, K, V>
-  implements Omit<IEnumItems<T, K, V>, typeof IS_ENUM_ITEMS>
+  implements Omit<IEnumItems<T, K, V>, typeof IS_ENUM_ITEMS | typeof ITEMS | typeof KEYS | typeof VALUES | 'labels'>
 {
   private __options__: EnumItemOptions | undefined;
   readonly items!: EnumItemsArray<T, K, V>;
   readonly keys!: K[];
+
+  constructor(init: T = {} as T, options?: EnumItemOptions) {
+    super();
+    // Do not use class field here, because don't want print this field in Node.js
+    Object.defineProperty(this, '__options__', {
+      value: options,
+      writable: false,
+      enumerable: false,
+      configurable: false,
+    });
+
+    const keys = Object.keys(init) as K[];
+    // Generate enum items array
+    const items = new EnumItemsArray<T, K, V>(init, options);
+    Object.freeze(items);
+    // @ts-expect-error: because use ITEMS to avoid naming conflicts in case of 'items' field name is taken
+    this[keys.includes('items') ? ITEMS : 'items'] = items;
+
+    // @ts-expect-error: because use KEYS to avoid naming conflicts in case of 'keys' field name is taken
+    this[keys.includes('keys') ? KEYS : 'keys'] = items[KEYS];
+
+    // @ts-expect-error: because use VALUES to avoid naming conflicts in case of 'values' field name is taken
+    this[keys.includes('values') ? VALUES : 'values'] = items[VALUES];
+
+    // Add keys to the instance, allows picking enum values by keys
+    items.forEach((item) => {
+      // @ts-expect-error: because of dynamic property
+      this[item.key] = item.value;
+    });
+
+    // @ts-expect-error: because use LABELS to avoid naming conflicts in case of 'labels' field name is taken
+    Object.defineProperty(this, keys.includes('labels') ? LABELS : 'labels', {
+      enumerable: true,
+      configurable: false,
+      get: () => items.labels,
+    });
+
+    Object.freeze(this);
+  }
   /**
    * - **EN:** A boolean value indicates that this is an enum collection instance.
    * - **CN:** 布尔值，表示这是一个枚举集合实例
@@ -75,66 +111,6 @@ export class EnumCollectionClass<
       return localize(this.__options__?.name);
     }
     return this.__options__?.name;
-  }
-
-  constructor(init: T = {} as T, options?: EnumItemOptions) {
-    super();
-    // Do not use class field here, because don't want print this field in Node.js
-    Object.defineProperty(this, '__options__', {
-      value: options,
-      writable: false,
-      enumerable: false,
-      configurable: false,
-    });
-
-    // Generate keys array
-    // exclude number keys with a "reverse mapping" value, it means those "reverse mapping" keys of number enums
-    const keys = Object.keys(init).filter(
-      (k) => !(/^-?\d+$/.test(k) && k === `${init[init[k as K] as K] ?? ''}`)
-    ) as K[];
-    const parsed = keys.map((key) => parseEnumItem<EnumItemInit<V>, K, V>(init[key], key));
-    keys.forEach((key, index) => {
-      const { value } = parsed[index];
-      // @ts-expect-error: because of dynamic property
-      this[key] = value;
-    });
-    Object.freeze(keys);
-    // @ts-expect-error: because use KEYS to avoid naming conflicts in case of 'keys' field name is taken
-    this[Object.keys(init).some((k) => k === 'keys') ? KEYS : 'keys'] = keys;
-
-    // Generate values array
-    const values = parsed.map((item) => item.value);
-    Object.freeze(values);
-    // @ts-expect-error: because use VALUES to avoid naming conflicts in case of 'values' field name is taken
-    this[Object.keys(init).some((k) => k === 'values') ? VALUES : 'values'] = values;
-
-    // Generate enum items array
-    const items = new EnumItemsArray<T, K, V>(
-      init,
-      this.__options__,
-      ...keys.map((key, index) => {
-        const { value, label } = parsed[index];
-        return new EnumItemClass<T[K], K, V>(key, value, label, init[key], this.__options__);
-      })
-    );
-    Object.freeze(items);
-    // @ts-expect-error: because use ITEMS to avoid naming conflicts in case of 'items' field name is taken
-    this[Object.keys(init).some((k) => k === 'items') ? ITEMS : 'items'] = items;
-
-    // Generate labels array
-    Object.defineProperty(this, Object.keys(init).some((k) => k === 'labels') ? LABELS : 'labels', {
-      enumerable: true,
-      configurable: false,
-      get: () => {
-        // Cannot save to static array because labels may be localized contents
-        // Should not use `items` in the closure because the getter function cannot be fully serialized
-        return ((this as unknown as { [ITEMS]: EnumItemsArray<T, K, V> })[ITEMS] ?? this.items).map(
-          (item) => item.label
-        );
-      },
-    });
-
-    Object.freeze(this);
   }
 
   label<KV extends V | K | NonNullable<PrimitiveOf<V>> | NonNullable<PrimitiveOf<K>> | undefined>(keyOrValue: KV) {
@@ -213,51 +189,4 @@ export class EnumCollectionClass<
   get rawType() {
     return this.items.rawType;
   }
-}
-
-function parseEnumItem<
-  T extends EnumItemInit<V>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  K extends EnumKey<any>,
-  V extends EnumValue,
->(init: T, key: K): StandardEnumItemInit<V> {
-  let value: V;
-  let label: string;
-  if (init != null) {
-    if (typeof init === 'number' || typeof init === 'string' || typeof init === 'symbol') {
-      value = init as V;
-      label = key as string;
-    } else if (typeof init === 'object') {
-      // Initialize using object
-      if (Object.prototype.toString.call(init) === '[object Object]') {
-        if ('value' in init && Object.keys(init).some((k) => k === 'value')) {
-          // type of {value, label}
-          value = init.value ?? key;
-          if ('label' in init && Object.keys(init).some((k) => k === 'label')) {
-            label = init.label;
-          } else {
-            label = key as string;
-          }
-        } else if ('label' in init && Object.keys(init).some((k) => k === 'label')) {
-          // typeof {label}
-          value = key as unknown as V;
-          label = init.label ?? key;
-        } else {
-          // {} empty object
-          value = key as unknown as V;
-          label = key as string;
-        }
-      } else {
-        // Probably Date, RegExp and other primitive types
-        value = init as V;
-        label = key as string;
-      }
-    } else {
-      throw new Error(`Invalid enum item: ${JSON.stringify(init)}`);
-    }
-  } else {
-    value = key as unknown as V;
-    label = key as string;
-  }
-  return { value, label };
 }